pyproject.toml
[project]
name = "DiscordBot"
version = "3.0.1"
description = "A Bot for Discord"
urls.Repository = "https://github.com/ddc/DiscordBot"
urls.Homepage = "https://github.com/ddc/DiscordBot"
license = {text = "MIT"}
readme = "README.md"
authors = [
    {name = "Daniel Costa", email = "ddcsoftwares@proton.me"},
]
maintainers = [
    {name = "Daniel Costa"},
]
keywords = [
    "python", "python3", "python-3",
    "DiscordBot", "discord-bots", "bot",
    "discord-py", "discord"
]
classifiers = [
    "Topic :: Communications :: Chat",
    "Development Status :: 5 - Production/Stable",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.14",
    "Operating System :: OS Independent",
    "Environment :: Other Environment",
    "Intended Audience :: Developers",
    "Natural Language :: English",
]
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.4",
    "beautifulsoup4>=4.14.3",
    "better-profanity>=0.7.0",
    "ddcdatabases[postgres]>=3.0.10",
    "discord-py>=2.6.4",
    "gTTS>=2.5.4",
    "openai>=2.20.0",
    "PyNaCl>=1.6.2",
    "pythonLogs>=6.0.2",
]

[dependency-groups]
dev = [
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "testcontainers[postgres]>=4.14.1",
    "poethepoet>=0.41.0",
    "ruff>=0.15.0",
    "black>=26.1.0",
]

[tool.poe.tasks]
linter.shell = "uv run ruff check --fix . && uv run black ."
profile.sequence = ["linter", {shell = "uv run python -m cProfile -o cprofile_unit.prof -m pytest tests/unit --no-cov"}]
profile-integration.sequence = ["linter", {shell = "uv run python -m cProfile -o cprofile_integration.prof -m pytest tests/integration --no-cov"}]
test.sequence = ["linter", {shell = "uv run pytest"}]
test-integration.sequence = ["linter", {shell = "uv run pytest tests/integration --no-cov"}]
hadolint.shell = "docker run --rm -i -v $(pwd)/.hadolint.yaml:/.config/hadolint.yaml:ro hadolint/hadolint < Dockerfile"
test-docker.shell = "uv run pytest tests/docker -v --no-cov"
updatedev.sequence = ["linter", {shell = "uv lock --upgrade && uv sync --all-extras --group dev"}]
migration.shell = "uv run --frozen alembic upgrade head"

[tool.pytest.ini_options]
addopts = "-v --import-mode=importlib --cov --cov-report=term --cov-report=xml --junitxml=junit.xml"
junit_family = "legacy"
pythonpath = ["."]
testpaths = ["tests/unit"]
asyncio_mode = "strict"
asyncio_default_fixture_loop_scope = "function"
markers = [
    "integration: marks tests as integration tests",
    "docker: Docker and compose file tests",
]
[tool.coverage.run]
omit = [
    "tests/*",
    "*/__init__.py",
]

[tool.coverage.report]
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.black]
line-length = 120
skip-string-normalization = true

[tool.ruff]
line-length = 120
target-version = "py314"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP"]
ignore = ["E501", "E402", "UP046", "UP047"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"tests/**/*.py" = ["S101", "S105", "S106", "S311", "SLF001", "F841"]

[tool.ruff.lint.isort]
known-first-party = ["DiscordBot"]
force-sort-within-sections = false
from-first = false
no-sections = true
